requirements.txt
argparse
qrcode
Pillow
six